+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 * [Scripts] Tool-tip plugin adapted from Bootstrap plugin
 * [Scripts] Pop-overs plugin adapted from Bootstrap plugin
 * [Scripts] Transitions plugin adapted from Bootstrap plugin
+* [Infrastructure] Error (404) page for testbed
 
 ### Fixed